requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
django-filter==24.3

# ─── Admin Panel (Jazzmin) ────────────────────────────────────────────────────
django-jazzmin==3.0.1

# ─── API Documentation (Swagger / ReDoc) ─────────────────────────────────────
drf-yasg==1.21.8

# ─── Security & Environment ───────────────────────────────────────────────────
python-dotenv==1.0.1
django-environ==0.11.2

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# ─── Static Files & Production ────────────────────────────────────────────────
whitenoise==6.7.0
gunicorn==23.0.0

# ─── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.3

# ─── Utilities ────────────────────────────────────────────────────────────────
Pillow==10.3.0
